Chore: Tidy up, split up files in main package
