+++ cmd/HellPot/HellPot_test.go
@@ -4,7 +4,26 @@ import (
 	"os"
 	"testing"
 	"time"
+
+	"github.com/yunginnanet/HellPot/internal/http"
 )
+
+func testMain(t interface{ Error(...any) }) (string, string, chan os.Signal, error) {
+	stopChan := make(chan os.Signal, 1)
+	log, logFile, resolvedConf, err := setup(stopChan)
+	if err == nil {
+		log.Info().Msg("🔥 Starting HellPot 🔥")
+		go func() {
+			terr := http.Serve(runningConfig)
+			if terr != nil {
+				t.Error("failed to serve HTTP: " + terr.Error())
+				close(stopChan)
+			}
+		}()
+	}
+	//goland:noinspection GoNilness
+	return resolvedConf, logFile, stopChan, err
+}
 
 func TestHellPot(t *testing.T) {
 	resolvedConf, logFile, stopChan, err := testMain(t)
